+++ docs/source/conf.py
@@ -1,28 +1,19 @@
-# Configuration file for the Sphinx documentation builder.
-#
-# This file only contains a selection of the most common options. For a full
-# list see the documentation:
-# https://www.sphinx-doc.org/en/master/usage/configuration.html
-
-# -- Path setup --------------------------------------------------------------
-
-# If extensions (or modules to document with autodoc) are in another directory,
-# add these directories to sys.path here. If the directory is relative to the
-# documentation root, use os.path.abspath to make it absolute, like shown here.
-#
-# import os
-# import sys
-# sys.path.insert(0, os.path.abspath('.'))
-
+import os
+import sys
+sys.path.insert(0, os.path.abspath('./../../'))
+import gamornet
 
 # -- Project information -----------------------------------------------------
 
-project = 'GaMorNet'
-copyright = '2020, Aritra Ghosh'
-author = 'Aritra Ghosh'
+project = 'gamornet'
+copyright = gamornet.__copyright__
+author = gamornet.__author__
 
 # The full version, including alpha/beta/rc tags
-release = '1.0'
+version = gamornet.__version__
+release = gamornet.__version__
+
+rst_epilog = '.. |copyright| replace:: %s' % copyright
 
 
 # -- General configuration ---------------------------------------------------
@@ -30,7 +21,11 @@ release = '1.0'
 # Add any Sphinx extension module names here, as strings. They can be
 # extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
 # ones.
-extensions = ['recommonmark']
+extensions = [
+    "sphinx.ext.autodoc",
+    "sphinx.ext.mathjax",
+    "sphinx.ext.napoleon",
+]
 
 # Add any paths that contain templates here, relative to this directory.
 templates_path = ['_templates']
@@ -38,7 +33,7 @@ templates_path = ['_templates']
 # List of patterns, relative to source directory, that match files and
 # directories to ignore when looking for source files.
 # This pattern also affects html_static_path and html_extra_path.
-exclude_patterns = []
+exclude_patterns = ['recommonmark']
 
 
 # -- Options for HTML output -------------------------------------------------
@@ -53,5 +48,5 @@ html_theme = 'default'
 # so a file named "default.css" will overwrite the builtin "default.css".
 html_static_path = []
 
-### Tgins that need to be set this way for RTD Integration ##
+### Things that need to be set this way for RTD Integration ##
 master_doc = 'index'