+++ gamornet/__init__.py
@@ -1,0 +1,5 @@
+
+__version__ = "0.3"
+__author__ = "Aritra Ghosh (http://ghosharitra.com)"
+__copyright__ = "Copyright 2020, Aritra Ghosh and Contributors"
+